+++ src/main/java/roomescape/controller/AdminThemeController.java
@@ -13,10 +13,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Theme;
-import roomescape.dto.CreateThemeRequest;
-import roomescape.dto.PopularThemeResponse;
-import roomescape.dto.ThemeReservationTimeResponse;
-import roomescape.dto.ThemeResponse;
+import roomescape.dto.theme.CreateThemeRequest;
+import roomescape.dto.theme.PopularThemeResponse;
+import roomescape.dto.theme.ThemeReservationTimeResponse;
+import roomescape.dto.theme.ThemeResponse;
 import roomescape.service.ThemeService;
 
 @RestController
