+++ src/main/java/roomescape/dto/theme/ThemeReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.theme;
 
 public record ThemeReservationTimeResponse(
         Long id,
