+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.CreateReservationTimeRequest;
+import roomescape.dto.reservationTime.CreateReservationTimeRequest;
 import roomescape.repository.ReservationTimeDao;
 
 @Service
