+++ src/main/java/roomescape/service/ThemeService.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.dto.CreateThemeRequest;
-import roomescape.dto.PopularThemeResponse;
-import roomescape.dto.ThemeReservationTimeResponse;
+import roomescape.dto.theme.CreateThemeRequest;
+import roomescape.dto.theme.PopularThemeResponse;
+import roomescape.dto.theme.ThemeReservationTimeResponse;
 import roomescape.repository.ReservationDao;
 import roomescape.repository.ReservationTimeDao;
 import roomescape.repository.ThemeDao;
